% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_clustering.R
\name{gibbs_sampling}
\alias{gibbs_sampling}
\title{Gibbs sampling}
\usage{
gibbs_sampling(data, k, class_labels, fix_vec, d = NULL, N = NULL,
  num_iter = NULL, burn = 0, mu_0 = NULL, df_0 = NULL,
  scale_0 = NULL, lambda_0 = 0.01, concentration_0 = 0.1,
  thinning = 25, outlier = FALSE, t_df = 4,
  record_posteriors = FALSE)
}
\arguments{
\item{data}{A matrix of the data being analysed.}

\item{k}{The number of clusters.}

\item{class_labels}{A vector of unsigned integers representing the initial 
cluster of the corresponding point in data}

\item{num_iter}{The number of iterations to sample over.}

\item{burn}{The number of iterations to record after (i.e. the burn-in).}

\item{mu_0}{A d-vector; prior on mean. If NULL defaults to mean of data.}

\item{df_0}{The prior on the degrees of freedom. if NULL defaults to d + 2.}

\item{scale_0}{The prior on the scale for the Inverse Wishart. If NULL 
generated using an empirical method.}

\item{lambda_0}{The prior of shrinkage for mean distribution.}

\item{concentration_0}{The prior for dirichlet distribution of cluster 
weights.}

\item{thinning}{The step between iterations for which results are recorded in
the mcmc output.}

\item{outlier}{A bool instructing the sampler to consider an additional 
outlier cluster following a t-distribution}

\item{t_df}{The degrees of freedom for the outlier t-distribution (default 
is 4)}

\item{record_posteriors}{A bool instructing the mcmc function to record the 
posterior distributions of the mean and variance for each cluster
(default is FALSE)}
}
\description{
Carries out gibbs sampling of data and returns a similarity matrix for points
}
