% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_clustering.R, R/normal_clustering.R
\name{phi_prior}
\alias{phi_prior}
\title{Generates a prior for the phi vector for each variable for the Dirichlet 
distribution}
\usage{
phi_prior(matrix_data)

phi_prior(matrix_data)
}
\arguments{
\item{matrix_data}{A matrix of data}

\item{matrix_data}{A matrix of data}
}
\value{
A list of vectors of the proportion of each level across all of 
matrix_data.

A list of vectors of the proportion of each level across all of 
matrix_data.
}
\description{
Generates a prior for the phi vector for each variable for the Dirichlet 
distribution
}
